pyproject.toml
[tool.poetry]
package-mode = false
name = "pyQuestionnaire"
version = "0.9.0-beta1"
description = ""
#authors = ["Apostolis Triantafyllou <apostolis@apostolistriantafyllou.gr>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
openpyxl = "^3.1.2"
python-sane = "^2.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
